Feature: Send switch on / off functions added
send_switch_on and send_switch_off have both been added and fully
tested.

Both check the device to send to against the current list of devices and
will return false if you are trying to switch a device which isn't on
the list. Therefore, the list of devices will need to be kept up to
date.
	new file:   README.md
	modified:   src/canbus_controller/canbus_controller.c
	modified:   src/canbus_controller/canbus_controller.h
	modified:   test/canbus_controller/test_canbus_controller.c